% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipal-package.R
\docType{package}
\name{clipal-package}
\alias{clipal}
\alias{clipal-package}
\title{clipal: Convert to cli errors}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

There have been many interfaces to raise errors with R over the years. A couple years ago, the tidyteam converged on using the 'cli' package and began the cumbersome process of rewriting erroring code. This task is a difficult one to automate, as there are all sorts of weird edge cases and stylistic divergences; "every unhappy family is unhappy in its own way." This RStudio add-in helps you convert your R code to use the 'cli' package for error messages using LLMs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/clipal}
  \item Report bugs at \url{https://github.com/simonpcouch/clipal/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
