% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipal.R
\name{clipal}
\alias{clipal}
\title{Create a cli pal}
\usage{
clipal(fn = "new_chat_claude", ..., .ns = "elmer")
}
\arguments{
\item{fn}{A \verb{new_*()} function, like from the elmer package.}

\item{...}{Additional arguments to \code{fn}.}

\item{.ns}{The package that the \verb{new_*()} function is exported from.}
}
\description{
A light wrapper around elmer chat functions to create a chat object with
a custom system prompt.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# to create a chat with claude:
clipal()

# or with OpenAI's o1-mini:
clipal(
  "new_chat_openai",
  model = "gpt-4o-mini"
)
\dontshow{\}) # examplesIf}
}
