% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_cli.R
\name{convert_to_cli}
\alias{convert_to_cli}
\title{Convert erroring code to use cli}
\usage{
convert_to_cli(x, clipal)
}
\arguments{
\item{x}{Lines of code that raise an error, as a character vector.}

\item{chat}{An elmer chat. Note that the system prompt will be overwritten.}
}
\description{
Convert erroring code to use cli
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clipal <- new_clipal()

convert_to_cli('stop("An error message.")', clipal)
\dontshow{\}) # examplesIf}
}
