% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipal-package.R
\docType{package}
\name{clipal-package}
\alias{clipal-package}
\title{clipal: Convert to cli errors}
\description{
RStudio add-in that helps you convert your R code to use the 'cli' package for error messages using LLMs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/clipal}
  \item Report bugs at \url{https://github.com/simonpcouch/clipal/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
