% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_cli.R
\name{convert_to_cli}
\alias{convert_to_cli}
\title{Convert erroring code to use cli}
\usage{
convert_to_cli(expr, cli_pal = .last_cli_pal)
}
\arguments{
\item{expr}{Lines of code that raise an error, as an expression.}

\item{cli_pal}{A cli pal created with \code{\link[=cli_pal]{cli_pal()}}.}
}
\description{
Convert erroring code to use cli
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cli_pal <- cli_pal()

convert_to_cli(stop("An error message."))
\dontshow{\}) # examplesIf}
}
